fix(harness): never auto-retry mutations on ENV_FAILURE
Root cause of the "nameID already taken" nightly failures (subspace,
subsubspace, entitlements): graphqlErrorWrapper retried ANY request on an
ENV_FAILURE connection reset. For a create, the server often commits
before the ~5s reset, so the retry re-issues the same nameID and the
server rejects the duplicate — failing a test for an entity that actually
exists. The nightly DB is recreated per run, so this is purely a
within-run retry-after-commit race, not stale data.

Fix at the source: move the ENV_FAILURE retry into the generated SDK
wrapper (getGraphqlClient), where the operation type is known, and retry
ONLY idempotent queries. Mutations run exactly once; a reset during a
mutation now surfaces as a labelled ENV_FAILURE (environmental,
server#6258) instead of a duplicate. This needs no per-helper changes and
covers every suite uniformly. graphqlErrorWrapper keeps only its
error-shaping role (no retry loop).

Consequently revert createSubspaceOrFail to plain de-masking: with the
retry gone there is no committed-then-retried create to recover, and a
genuine `already taken` should surface rather than be silently looked up.
The de-masking, round-trip assertions, and tightened negatives from
earlier in this PR are unchanged.
