+++ lib/src/utils/graphql.wrapper.ts
@@ -38,23 +38,17 @@ const classifyNonGraphqlError = (
 };
 
 /**
- * Retry only ENV_FAILURE (connection-level) failures — never GraphQL errors,
- * which are legitimate API responses to assert on. These come from a ~5s
- * connection reset by a network element between the runner and the cluster when
- * a heavy request (e.g. subspace-tree creation) hasn't responded yet
- * (test-suites#563). A retry gives a fresh connection a chance to complete under
- * the cutoff. The durable fix is server-side latency (alkem-io/server#6258);
- * this is resilience until then. Caveat: for a mutation the server may have
- * committed before the reset, so a retry can hit a duplicate/conflict — for the
- * ephemeral, unique-named test data here that is harmless (it re-fails at worst,
- * never silently corrupts a real store).
+ * Shapes thrown request errors into the `{ error: { errors } }` contract the
+ * helpers/tests rely on. It does NOT retry.
+ *
+ * ENV_FAILURE (connection-level) retries live one layer deeper, in the SDK
+ * wrapper (`getGraphqlClient`), where the operation type is known — so only
+ * idempotent QUERIES are retried. Mutations must never be auto-retried: the
+ * server may have committed a create before a ~5s connection reset, and
+ * re-issuing it hits `nameID already taken`, failing a test for an entity that
+ * actually exists (the 2026-07-15 nightly cascades). By the time we catch here,
+ * any query retries have already happened; we just classify and report.
  */
-const ENV_FAILURE_MAX_ATTEMPTS = 3;
-const ENV_FAILURE_RETRY_BASE_MS = 1000;
-
-const sleep = (ms: number): Promise<void> =>
-  new Promise((resolve) => setTimeout(resolve, ms));
-
 export const graphqlErrorWrapper = async <TData>(
   fn: (authToken: string | undefined) => GraphQLReturnType<TData>,
   userRole?: TestUser,
@@ -64,42 +58,31 @@ export const graphqlErrorWrapper = async <TData>(
     const userModel = TestUserManager.getUserModelByType(userRole);
     authToken = userModel.authToken;
   }
-  const wrapperStartedAt = Date.now();
-  for (let attempt = 1; ; attempt++) {
-    const startedAt = Date.now();
-    try {
-      LogManager.getLogger().info(`Executing request: ${fn}`);
-      return await fn(authToken);
-    } catch (error) {
-      const err = error as ErrorType;
-      if (!err.response || !err.response.errors) {
-        const elapsedMs = Date.now() - startedAt;
-        const { code, detail } = classifyNonGraphqlError(error);
-        if (code === "ENV_FAILURE" && attempt < ENV_FAILURE_MAX_ATTEMPTS) {
-          const delay = ENV_FAILURE_RETRY_BASE_MS * attempt;
-          LogManager.getLogger().warn(
-            `[ENV_FAILURE] request failed after ${elapsedMs}ms (attempt ${attempt}/${ENV_FAILURE_MAX_ATTEMPTS}); retrying in ${delay}ms: '${fn}'`,
-          );
-          await sleep(delay);
-          continue;
-        }
-        const totalMs = Date.now() - wrapperStartedAt;
-        LogManager.getLogger().error(
-          `[${code}] request failed on attempt ${attempt} (${elapsedMs}ms this attempt, ${totalMs}ms total incl. retries): '${fn}'`,
-        );
-        LogManager.getLogger().error("Returned error:");
-        LogManager.getLogger().error(err);
-        return {
-          error: {
-            errors: [
-              {
-                message: `[${code}] ${detail} (after ${totalMs}ms across ${attempt} attempt(s))`,
-                code,
-              },
-            ],
-          },
-        };
-      } else {
+  const startedAt = Date.now();
+  try {
+    LogManager.getLogger().info(`Executing request: ${fn}`);
+    return await fn(authToken);
+  } catch (error) {
+    const err = error as ErrorType;
+    if (!err.response || !err.response.errors) {
+      const elapsedMs = Date.now() - startedAt;
+      const { code, detail } = classifyNonGraphqlError(error);
+      LogManager.getLogger().error(
+        `[${code}] request failed (${elapsedMs}ms): '${fn}'`,
+      );
+      LogManager.getLogger().error("Returned error:");
+      LogManager.getLogger().error(err);
+      return {
+        error: {
+          errors: [
+            {
+              message: `[${code}] ${detail} (after ${elapsedMs}ms)`,
+              code,
+            },
+          ],
+        },
+      };
+    } else {
       const badErrors = err.response.errors.filter(
         (e) =>
           e.extensions.code !== "BAD_USER_INPUT" &&
@@ -120,7 +103,6 @@ export const graphqlErrorWrapper = async <TData>(
           })),
         },
       };
-      }
     }
   }
 };
