+++ server-api/src/functional-api/journey/subspace/subspace.request.params.ts
@@ -63,29 +63,20 @@ export const createSubspace = async (
 };
 
 /**
- * Setup-time subspace creation that is resilient to the ENV_FAILURE
- * retry-after-commit race, and surfaces genuine failures instead of masking
- * them.
+ * Setup-time subspace creation that surfaces failures instead of masking them.
  *
- * Two problems this replaces the plain
- * `createSubspace(...).data?.createSubspace.id ?? ''` idiom for:
+ * The `createSubspace(...).data?.createSubspace.id ?? ''` idiom used in test
+ * setup turns a failed create into an empty string; later mutations/queries then
+ * reject that empty id with a misleading `UUID not valid:`. That is how the
+ * 2026-07-15 nightly produced a 12-failure cascade across
+ * subspace-sorting-and-pinning and subsubspace with no trace of the real cause.
+ * Throw the actual error at the true failure point instead (same de-masking as
+ * `createRootSpace`, test-suites#577 / #563).
  *
- * 1. Masking. The `?? ''` turns a failed create into an empty string; later
- *    mutations/queries then reject that empty id with a misleading
- *    `UUID not valid:`. That is how the 2026-07-15 nightly produced a 12-failure
- *    cascade across subspace-sorting-and-pinning and subsubspace with no trace
- *    of the real cause. We throw the actual error at the true failure point
- *    instead (same de-masking as `createRootSpace`, test-suites#577 / #563).
- *
- * 2. The retry-after-commit race. On a ~5s connection reset (test-suites#563 /
- *    server#6258), `graphqlErrorWrapper` retries the mutation — but a create may
- *    have already committed server-side before the reset, so the retry comes
- *    back as `nameID already taken` (BAD_USER_INPUT). The subspace *does* exist;
- *    failing setup for a create that actually succeeded is wrong. Because the
- *    nameID is unique per run, `already taken` here reliably means our own
- *    committed-then-retried create — so we look it up by nameID under the parent
- *    and return it. (This is the caveat the wrapper's retry comment calls
- *    "harmless"; for a create it is not — hence this recovery.)
+ * Note: the earlier `nameID already taken` failures were the wrapper retrying a
+ * create that had already committed before a connection reset. That is now fixed
+ * at the source — the SDK wrapper (`getGraphqlClient`) never retries mutations —
+ * so `already taken` here again means a genuine duplicate and should surface.
  */
 export const createSubspaceOrFail = async (
   subspaceName: string,
@@ -100,33 +91,15 @@ export const createSubspaceOrFail = async (
     userRole
   );
   const id = response.data?.createSubspace?.id;
-  if (id) {
-    return id;
+  if (!id) {
+    const detail = response.error
+      ? JSON.stringify(response.error.errors)
+      : 'server returned no createSubspace.id and no error';
+    throw new Error(
+      `Failed to create subspace '${subspaceName}' (nameID '${subspaceNameId}', parent '${parentId}'): ${detail}`
+    );
   }
-
-  const alreadyTaken = response.error?.errors?.some(e =>
-    String((e as { message?: unknown }).message ?? '').includes(
-      'nameID is already taken'
-    )
-  );
-  if (alreadyTaken) {
-    const existing = await getSubspacesData(parentId);
-    const subspaces = (existing.data?.lookup?.space?.subspaces ?? []) as Array<{
-      id: string;
-      nameID: string;
-    }>;
-    const match = subspaces.find(s => s.nameID === subspaceNameId);
-    if (match?.id) {
-      return match.id;
-    }
-  }
-
-  const detail = response.error
-    ? JSON.stringify(response.error.errors)
-    : 'server returned no createSubspace.id and no error';
-  throw new Error(
-    `Failed to create subspace '${subspaceName}' (nameID '${subspaceNameId}', parent '${parentId}'): ${detail}`
-  );
+  return id;
 };
 
 export const subspaceVariablesData = (
